update the gmodel tools

+++ extends/gmodel/utils/DBFactory.class.php
@@ -3,7 +3,7 @@
 namespace gmodel\utils;
 
 /**
- * 数据库，数据表。
+ * 创建数据库，数据表
  * @author yangjian<yangjian102621@gmail.com>
  *
  */
@@ -16,6 +16,10 @@ class DBFactory {
      */
     private static $DB_CONN = null; //数据库连接
 
+    /**
+     * 默认值映射
+     * @var array
+     */
     private static $DEFAULT_VALUE_KEYWORD = array(
         "CURRENT_TIMESTAMP", "NULL"
     );
@@ -38,8 +42,8 @@ class DBFactory {
             "dbuser" => $root->getAttribute("dbuser"),
             "dbpass" => $root->getAttribute("dbpass"),
             "dbname" => $root->getAttribute("dbname"),
-            "charset" => $root->getAttribute("charset"),
-            "table-prefix" => $root->getAttribute("table-prefix")
+            "charset" => $root->getAttribute("charset")
+//            "table-prefix" => $root->getAttribute("table-prefix")
         );
 
         self::$DB_CONN = mysqli_connect($configs["dbhost"], $configs["dbuser"], $configs["dbpass"]);
@@ -61,20 +65,24 @@ class DBFactory {
 
         //生成数据库配置文档
         $dbConfigFile = APP_PATH."configs/db.config.php";
-        $tempContent = file_get_contents(dirname(__DIR__)."/template/db.config.tpl");
-        if ( $tempContent != "" ) {
-            $content = str_replace("{db_host}", $configs["dbhost"], $tempContent);
-            $content = str_replace("{db_user}", $configs["dbuser"], $content);
-            $content = str_replace("{db_name}", $configs["dbname"], $content);
-            $content = str_replace("{db_pass}", $configs["dbpass"], $content);
-            $content = str_replace("{db_charset}", $configs["charset"], $content);
-            $content = str_replace("{table_prefix}", $configs["table-prefix"], $content);
+        if ( !file_exists($dbConfigFile) ) {
+            $tempContent = file_get_contents(dirname(__DIR__)."/template/db.config.tpl");
+            if ( $tempContent != "" ) {
+                $content = str_replace("{db_host}", $configs["dbhost"], $tempContent);
+                $content = str_replace("{db_user}", $configs["dbuser"], $content);
+                $content = str_replace("{db_name}", $configs["dbname"], $content);
+                $content = str_replace("{db_pass}", $configs["dbpass"], $content);
+                $content = str_replace("{db_charset}", $configs["charset"], $content);
+                $content = str_replace("{table_prefix}", $configs["table-prefix"], $content);
 
-            if ( file_put_contents($dbConfigFile, $content) !== false ) {
-                tprintOk("create db config file '{$dbConfigFile}' successfully！");
-            } else {
-                tprintError("create db config file '{$dbConfigFile}' faild!");
+                if ( file_put_contents($dbConfigFile, $content) !== false ) {
+                    tprintOk("create db config file '{$dbConfigFile}' successfully！");
+                } else {
+                    tprintError("create db config file '{$dbConfigFile}' faild!");
+                }
             }
+        } else {
+            tprintWarning("config file '{$dbConfigFile}' has existed, skiped.");
         }
 
         self::createTables($configs);
@@ -112,14 +120,14 @@ class DBFactory {
             $fields = $value->find("fields", 0);
             if ( $fields ) {
                 foreach( $fields->children() as $fd ) {
-                    if ( $fd->default || $fd->default === "0" ) {   //has default value
+                    if ( $fd->default || $fd->default === "0" || $fd->default === '' ) {   //has default value
                         if ( in_array($fd->default, self::$DEFAULT_VALUE_KEYWORD) ) {
-                            $sql .= "`{$fd->name}` {$fd->type} NOT NULL DEFAULT {$fd->default} COMMENT '{$fd->comment}',";
+                            $sql .= "`{$fd->name}` {$fd->type} DEFAULT {$fd->default} COMMENT '{$fd->comment}',";
                         } else {
-                            $sql .= "`{$fd->name}` {$fd->type} NOT NULL DEFAULT '{$fd->default}' COMMENT '{$fd->comment}',";
+                            $sql .= "`{$fd->name}` {$fd->type} DEFAULT '{$fd->default}' COMMENT '{$fd->comment}',";
                         }
                     } else { //has not default value
-                        $sql .= "`{$fd->name}` {$fd->type} NOT NULL COMMENT '{$fd->comment}',";
+                        $sql .= "`{$fd->name}` {$fd->type} DEFAULT NULL COMMENT '{$fd->comment}',";
                     }
 
                     //创建索引
